Allow olly to attach to an external process
- The command line changes are nontrivial in that they cannot be
expressed with usual `Cmdliner` combinators. Attaching is done with an
`--attach` option that is incompatible with the normal `EXECUTABLE`
arguments

- `EXECUTABLE` is now a sequence of arguments, rather than a single
space-separated one. This makes the above mutual exclusion slightly
easier to represent, and is more in-line with existing tools (`perf`,
`gdb --args`).

- `Unix.kill pid 0` is used to check if the process is alive, (since
`wait` can only be used for child processes), this is valid and
intended usage according to the POSIX specification.